feat(text search): adding additional config options for lower case and stripping symbols
I've noticed that I have some really odd named artists in my lidarr instance that come with very
specific capitalization and symbols. Knowing how the lidarr cache is case and symbol sensitive, it
might make sense to give the user the option to force lowercase and remove symbols from the artist
names when building the text search cache. I've run this against my database and it seems to be
working well, but if you don't want it, I'll keep it in my fork.